+++ include/roq/api.hpp
@@ -10,6 +10,7 @@
 
 // enums
 
+#include "roq/buffer_capacity.hpp"
 #include "roq/connection_status.hpp"
 #include "roq/decimals.hpp"
 #include "roq/encoding.hpp"