+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Add notional fields to `ReferenceData` (#252)
+* Add new fields to `RateLimitTrigger` (#251)
+
 ## 0.8.6 &ndash; 2022-07-18
 
 ### Fixed